feat: - Addressed reviewers comments - renamed from SwitchLowering to SwitchPeephole -Flattened logic to use continue and removed indenting - Added test case for case 0, 1 and default
